claude_evolve/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-evolve"
version = "0.1.0"
description = "Evolutionary code/artifact optimization for Claude Code - extracted from OpenEvolve"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "dacite>=1.8",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
claude-evolve = "claude_evolve.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
